+++ src/elements/Asset.php
@@ -2316,33 +2316,34 @@ class Asset extends Element
 
         $tempFilePath = FileHelper::normalizePath($tempFilePath);
 
-        // Is it within one of our temp directories?
+        // Make sure it's within a known temp path, the project root, or storage/ folder
         $pathService = Craft::$app->getPath();
-        $tempDirs = [
-            $this->_normalizeTempPath($pathService->getTempPath()),
-            $this->_normalizeTempPath($pathService->getTempAssetUploadsPath()),
-            $this->_normalizeTempPath(sys_get_temp_dir()),
+        $allowedRoots = [
+            [$pathService->getTempPath(), true],
+            [$pathService->getTempAssetUploadsPath(), true],
+            [sys_get_temp_dir(), true],
+            [Craft::getAlias('@root'), false],
+            [Craft::getAlias('@storage'), false],
         ];
 
-        $tempDirs = array_filter($tempDirs, function($value) {
-            return ($value !== false);
-        });
-
-        foreach ($tempDirs as $allowedFolder) {
-            if (StringHelper::startsWith($tempFilePath, $allowedFolder)) {
-                return true;
+        $inAllowedRoot = false;
+        foreach ($allowedRoots as [$root, $isTempDir]) {
+            $root = $this->_normalizeTempPath($root);
+            if ($root !== false && StringHelper::startsWith($tempFilePath, $root)) {
+                // If this is a known temp dir, we’re good here
+                if ($isTempDir) {
+                    return true;
+                }
+                $inAllowedRoot = true;
+                break;
             }
         }
-
-        // Make sure it's within the project root somewhere
-        $projectRoot = $this->_normalizeTempPath(Craft::getAlias('@root'));
-        if (!StringHelper::startsWith($tempFilePath, $projectRoot)) {
+        if (!$inAllowedRoot) {
             return false;
         }
 
-        // Make sure it's not within a system directory
+        // Make sure it's *not* within a system directory though
         $systemDirs = $pathService->getSystemPaths();
-
         $systemDirs = array_map([$this, '_normalizeTempPath'], $systemDirs);
         $systemDirs = array_filter($systemDirs, function($value) {
             return ($value !== false);